Add schedule trigger to workflows
The repository's GitHub Actions workflows are configured to run on any push or pull request that changes relevant files.
If a problem was introduced, the workflow run will fail.

Workflow run failures might also result from external changes (e.g., an update to a tool used by the workflow or the
GitHub actions runner environment). It is important for this breakage to be detected soon after it occurs so that the
project maintainer has time to fix the infrastructure before the problem impacts project development work. This is done
by also adding a schedule trigger to the workflows so that they will run periodically.