Always use Go version specified by `use` argument
The `use` verb adds the path of the binary folder for the Go version specified by the user to the system `PATH`.

It obtains the binary folder path from the output of the `go env GOROOT` command, using the `go` command for the
specified version.

Starting from Go 1.21.0, the Go toolchain is decoupled from the Go command. When a `go` command is invoked from the path
of a Go module, if the `go` or `toolchain` directives in the module's `go.mod` file specify a version higher than the
version of the command then the toolchain version indicated by the directive will be used instead of the version
matching the command version.

The `go env GOROOT` command returns the path of the toolchain, not of the command. This means that `gvm use` could
configure a different version of the Go command for use than the one specified by the user.

For example, with a module that has `go` directive value `1.23.6`, this happens when :

```
$ gvm use go1.23.5

$ which go

/c/Users/foo/sdk/go1.23.6/bin/go
```

(note that I specified that 1.23.5 should be used, but 1.23.6 was used instead)

Even though the selected version was appropriate for use with the module, it is still essential that the tool honor the
user's wishes (for example, they might be intentionally using an inappropriate version to perform an experiment). Go
will handle notifying the user if the version they selected is inappropriate.

This unexpected behavior can be avoided by setting the `GOTOOLCHAIN` environment variable before running the `go env
GOROOT` command. `GOTOOLCHAIN` overrides the automagical toolchain version selection, which forces `go env GOROOT` to
return the path matching the specified version, and thus for the `use` verb to always behave as the user intends:

```
$ gvm use go1.23.5

$ which go

/c/Users/foo/sdk/go1.23.5/bin/go
```

The `GOTOOLCHAIN` environment variable is set in a subshell, so this does not affect the functioning of Go for the user.
Go will automagically select the toolchain version just as before (the `use` verb is only intended to control the
command version, not the toolchain version).